short direct

+++ src/Instructions/ld_ldi.c
@@ -6,7 +6,7 @@
 /*   By: jle-quer <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/25 19:21:53 by jle-quer          #+#    #+#             */
-/*   Updated: 2016/12/08 18:24:13 by lbaudran         ###   ########.fr       */
+/*   Updated: 2016/12/09 14:51:23 by lbaudran         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,14 +55,14 @@ void	apply_ldi(t_data *data, t_list *elem, unsigned int *param_types,
 			param_types[1] == REG_CODE))
 	{
 	if (param_types[0] == DIR_CODE)
-		value_one = get_ind_size_at_add(data, elem, params[0]);
+		value_one = (short)get_ind_size_at_add(data, elem, params[0]);
 	else if (param_types[0] == IND_CODE)
 		value_one = recup_ind(data, (short)params[0], elem->pc);
 	else if (param_types[0] == REG_CODE)
 		value_one = get_ind_size_at_add(data, elem, elem->reg_number[params[0]]
 				);
 	if (param_types[1] == DIR_CODE)
-		value_two = params[1];
+		value_two = (short)params[1];
 	else if (param_types[1] == REG_CODE)
 		value_two = elem->reg_number[params[1]];
 	s = (value_one + value_two) % IDX_MOD;
